Create JMusicQuick
Methods to speed up the creation, display, and playing of music using the jMusic library. score methods are called by the play, notate, and write methods to create a score from at the very least an array of pitches (ints 0 to 127) and respectively play, display in music notation, or write a midi file of the music.